# 20201202/WA80/WA81/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WA80",
    "version": "1.0"
  },
  "paths": {
    "/api/Product": {
      "get": {
        "tags": [
          "Product"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/api/Product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/api/Product/Search/{filter}": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "filter",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ModelState": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "Category": {
        "required": [
          "categoryName",
          "description"
        ],
        "type": "object",
        "properties": {
          "rowVersion": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/ModelState"
          },
          "modifiedProperties": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryName": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string"
          },
          "picture": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "nullable": true
          },
          "pictureBase64": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Suppliers": {
        "required": [
          "companyName"
        ],
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "integer",
            "format": "int32"
          },
          "companyName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "contactName": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "contactTitle": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "address": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "city": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "region": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "country": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "phone": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "fax": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "homePage": {
            "type": "string",
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CustomerDemographics": {
        "type": "object",
        "properties": {
          "customerTypeId": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "customerDesc": {
            "type": "string",
            "nullable": true
          },
          "customerCustomerDemo": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerCustomerDemo"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CustomerCustomerDemo": {
        "type": "object",
        "properties": {
          "customerId": {
            "maxLength": 5,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "customerTypeId": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/Customers"
          },
          "customerType": {
            "$ref": "#/components/schemas/CustomerDemographics"
          }
        },
        "additionalProperties": false
      },
      "Customers": {
        "required": [
          "companyName"
        ],
        "type": "object",
        "properties": {
          "customerId": {
            "maxLength": 5,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "contactName": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "contactTitle": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "address": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "city": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "region": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "country": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "phone": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "fax": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "customerCustomerDemo": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerCustomerDemo"
            },
            "nullable": true
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Region": {
        "required": [
          "regionDescription"
        ],
        "type": "object",
        "properties": {
          "regionId": {
            "type": "integer",
            "format": "int32"
          },
          "regionDescription": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "territories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Territories"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Territories": {
        "required": [
          "territoryDescription"
        ],
        "type": "object",
        "properties": {
          "territoryId": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "territoryDescription": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "regionId": {
            "type": "integer",
            "format": "int32"
          },
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "employeeTerritories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeTerritories"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EmployeeTerritories": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "integer",
            "format": "int32"
          },
          "territoryId": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "employee": {
            "$ref": "#/components/schemas/Employees"
          },
          "territory": {
            "$ref": "#/components/schemas/Territories"
          }
        },
        "additionalProperties": false
      },
      "Employees": {
        "required": [
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "integer",
            "format": "int32"
          },
          "lastName": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "firstName": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string"
          },
          "title": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "titleOfCourtesy": {
            "maxLength": 25,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "hireDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "address": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "city": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "region": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "country": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "homePhone": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "extension": {
            "maxLength": 4,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "photo": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "reportsTo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "photoPath": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "reportsToNavigation": {
            "$ref": "#/components/schemas/Employees"
          },
          "employeeTerritories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeTerritories"
            },
            "nullable": true
          },
          "inverseReportsToNavigation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Employees"
            },
            "nullable": true
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Shippers": {
        "required": [
          "companyName"
        ],
        "type": "object",
        "properties": {
          "shipperId": {
            "type": "integer",
            "format": "int32"
          },
          "companyName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "phone": {
            "maxLength": 24,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Order": {
        "type": "object",
        "properties": {
          "rowVersion": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/ModelState"
          },
          "modifiedProperties": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "customerId": {
            "maxLength": 5,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "employeeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "orderDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "requiredDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shippedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shipVia": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "freight": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "shipName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "shipAddress": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "shipCity": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "shipRegion": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "shipPostalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "shipCountry": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/Customers"
          },
          "employee": {
            "$ref": "#/components/schemas/Employees"
          },
          "shipViaNavigation": {
            "$ref": "#/components/schemas/Shippers"
          },
          "orderDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDetail"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderDetail": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "discount": {
            "type": "number",
            "format": "float"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "additionalProperties": false
      },
      "Product": {
        "required": [
          "productName"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "productName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string"
          },
          "supplierId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "categoryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "quantityPerUnit": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "unitPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "unitsInStock": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "unitsOnOrder": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "reorderLevel": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "discontinued": {
            "type": "boolean"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "supplier": {
            "$ref": "#/components/schemas/Suppliers"
          },
          "orderDetails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDetail"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}